Print character only if printable.
When a character is not printable, the readline did still print it. In
some cases, it could create weird bug (see [1] or [2]). Since we have the
`Hoa\Ustring::isCharPrintable`, we can check whether a character is
printable or not. Consequently, here is the new logic:

  * We read a character,
  * The buffer receives this character,
  * If there is no mapping for this character and if it is not
    printable, then we do not print it. We do nothing with it.

So, if we would like to support a control character that is not
printable, we must associate a mapping to this character (like we do for
`Ctrl-A` for instance).

[1]: https://github.com/hoaproject/Console/issues/16
[2]: https://github.com/hoaproject/Console/issues/17