Create the `Input` interface.
Exactly like the `Output` class, the `Input` interface represents a
stream of the program, here the input stream. We then create the
`Console::setInput` and `Console::getInput` methods.

By default, if no `Input` instance is provided, an `Hoa\File\Read`
instance will be used representing `php://stdin`. If it has been closed
then `/dev/tty` will be used instead.

This way, we are ensured methods will behave as expected, whether or not
the standard input has been closed or not.