Introduce the `Output` object.
First we define an `Output` object. It represents the output of a
program, e.g. the standard output.

Second we define the `Console::setOutput` and `Console::getOutput`
methods to set or get the output object. Since it extends the
`Hoa\Stream\IStream\Out` interface, we can use any stream library as the
output, like:

    Hoa\Console\Console::setOutput(new Hoa\File\Write('/tmp/foo'));

In this case, all the outputs of the program are going to be manually
redirected to the `/tmp/foo` file.