Consider multiplexer while writing on the output.
Sometimes, when writing on the output, one might would like to by-pass
the multiplexer or to write a specific control sequence for the
multiplexer (if running). Dealing with all multiplexers and specific
vendor control sequences are hard. So we can tell the output to consider
the multiplexer and it will try to directly ensure its support. For
instance, for TMUX(1), it automatically write a control sequence to send
the original control sequence directly to the terminal, i.e. it
by-passes TMUX(1).