Console: Solve `stty -f` vs. `stty -F` issue.
On Linux, this is the `stty -F $file` or `stty --file $file` options.
On BSD, this is the `stty -f $file` option.

There is a conflict here and this is difficult to have something
cross-platform. Fortunately, both architectures use the standard input
as the `-F`, `--file` or `-f` value. So writing `stty < $file` seems to
fix the issue.